requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
cinemagoer==2022.12.27
click==8.1.3
continuous-threading==2.0.6
flatbuffers==23.3.3
frozenlist==1.3.3
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.54.0
gTTS==2.3.1
h5py==3.8.0
idna==3.4
IMDbPY==2022.7.9
jax==0.4.8
joblib==1.2.0
keras==2.12.0
libclang==16.0.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.1.0
multidict==6.0.4
numpy==1.23.5
oauthlib==3.2.2
openai==0.27.5
opt-einsum==3.3.0
packaging==23.1
Pillow==9.5.0
protobuf==4.22.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pydub==0.25.1
requests==2.29.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
soupsieve==2.4.1
SpeechRecognition==3.10.0
SQLAlchemy==2.0.11
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
threadpoolctl==3.1.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.3.2
wikipedia==1.4.0
wrapt==1.14.1
yarl==1.9.2